+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_params.py
@@ -371,6 +371,7 @@ def load_arguments(self, _):
     with self.argument_context('webapp ssh') as c:
         c.argument('slot', options_list=['--slot', '-s'], help='the name of the slot. Default to the productions slot if not specified')
         c.argument('timeout', options_list=['--timeout', '-t'], help='timeout in seconds. Defaults to none')
+        c.argument('skip_cert_check', options_list=['--skip-cert-check'], help='skip ssl certificate check', default=False)
 
     with self.argument_context('functionapp') as c:
         c.ignore('app_instance', 'slot')