+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/custom.py
@@ -2452,7 +2452,7 @@ def webapp_up(cmd, name, location=None, sku=None, dryrun=False, logs=False, laun
                           "%s -l %s", _url, name, location)
 
 
-def _ping_scm_site(cmd, resource_group, name):
+def _ping_scm_site(cmd, resource_group, name, skip_cert_check=False):
     #  wake up kudu, by making an SCM call
     import requests
     #  work around until the timeout limits issue for linux is investigated & fixed
@@ -2460,14 +2460,18 @@ def _ping_scm_site(cmd, resource_group, name):
     scm_url = _get_scm_url(cmd, resource_group, name)
     import urllib3
     authorization = urllib3.util.make_headers(basic_auth='{}:{}'.format(user_name, password))
-    requests.get(scm_url + '/api/settings', headers=authorization)
+    if skip_cert_check:
+        logger.info('skipping ssl check for %s', scm_url)
+        urllib3.disable_warnings()
+    requests.get(scm_url + '/api/settings', headers=authorization, verify=(not skip_cert_check))
 
 
 def is_webapp_up(tunnel_server):
     return tunnel_server.is_webapp_up()
 
 
-def create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=None, timeout=None):
+def create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=None, timeout=None,
+                              skip_cert_check=False):
     webapp = show_webapp(cmd, resource_group_name, name, slot)
     is_linux = webapp.reserved
     if not is_linux:
@@ -2487,7 +2491,7 @@ def create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=No
     scm_url = _get_scm_url(cmd, resource_group_name, name, slot)
 
     tunnel_server = TunnelServer('', port, scm_url, profile_user_name, profile_user_password)
-    _ping_scm_site(cmd, resource_group_name, name)
+    _ping_scm_site(cmd, resource_group_name, name, skip_cert_check)
 
     _wait_for_webapp(tunnel_server)
 
@@ -2553,12 +2557,13 @@ def _start_ssh_session(hostname, port, username, password):
         c.close()
 
 
-def ssh_webapp(cmd, resource_group_name, name, slot=None, timeout=None):  # pylint: disable=too-many-statements
+def ssh_webapp(cmd, resource_group_name, name, slot=None, timeout=None, skip_cert_check=False):  # pylint: disable=too-many-statements
     import platform
     if platform.system() == "Windows":
         raise CLIError('webapp ssh is only supported on linux and mac')
     else:
-        create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=slot, timeout=timeout)
+        create_tunnel_and_session(cmd, resource_group_name, name, port=None, slot=slot, timeout=timeout,
+                                  skip_cert_check=skip_cert_check)
 
 
 def create_devops_build(